requirements.txt
streamlit==1.39.0
pyjwt==2.9.0
extra-streamlit-components==0.1.71
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-dotenv==1.0.1
firebase-admin
pyrebase4
setuptools
